pyproject.toml
[tool.poetry]
name = "git-to-docs-platform"
version = "0.1.0"
description = "Zero-friction Git-to-docs platform that automatically generates beautiful documentation websites"
authors = ["Git-to-Docs Team <team@docify.dev>"]
readme = "README.md"
packages = [{include = "docify"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.104.1"
uvicorn = {extras = ["standard"], version = "^0.24.0"}
tree-sitter = "^0.20.4"
tree-sitter-python = "^0.20.4"
tree-sitter-javascript = "^0.20.1"
tree-sitter-typescript = "^0.20.3"
tree-sitter-java = "^0.20.2"
tree-sitter-go = "^0.20.0"
tree-sitter-rust = "^0.20.4"
tree-sitter-cpp = "^0.20.0"
gitpython = "^3.1.40"
pydantic = "^2.5.0"
pydantic-settings = "^2.1.0"
httpx = "^0.25.2"
openai = "^1.3.7"
google-generativeai = "^0.3.2"
google-genai = "^1.56.0"
jinja2 = "^3.1.2"
click = "^8.1.7"
python-multipart = "^0.0.6"
aiofiles = "^23.2.1"
psutil = "^5.9.6"
boto3 = "^1.34.0"
redis = "^5.0.1"
pillow = "^10.1.0"
matplotlib = "^3.8.2"
numpy = "^1.25.2"
pyjwt = "^2.8.0"
rich = "^13.7.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-asyncio = "^0.21.1"
hypothesis = "^6.92.1"
black = "^23.11.0"
isort = "^5.12.0"
flake8 = "^6.1.0"
mypy = "^1.7.1"
pre-commit = "^3.6.0"

[tool.poetry.scripts]
docify = "docify.cli:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py311']

[tool.isort]
profile = "black"
multi_line_output = 3

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
markers = [
    "property: marks tests as property-based tests (deselect with '-m \"not property\"')",
]